fixing coloring methods

+++ tplink-smartdevices/Data/BulbHSV.cs
@@ -3,8 +3,8 @@
     public class BulbHSV
     {
         public int Hue { get; set; }
-        public byte Saturation { get; set; }
-        public byte Value { get; set; }
+        public int Saturation { get; set; }
+        public int Value { get; set; }
         public int Brightness { get; set; }
     }
 }
